feat: Move most of file klondike_game.dart to klondike_world.dart.
